Correct package name in java, separate index.js
