+++ index.ios.js
@@ -1,0 +1,6 @@
+'use strict'
+
+import { NativeModules } from 'react-native';
+const { colorGrabber } = NativeModules;
+
+export default colorGrabber;
